# Pertemuan 11/Restfull API/public/assets/api.yaml
openapi: 3.0.0
paths:
...
  /creekConditions:
    get:
      tags:
        - CreekConditions
      summary: 'Find list CreekConditions'
      description: 'Returns list of CreekConditions'
      operationId: getCreekConditions
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/CreekConditions' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/CreekConditions' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/CreekConditions' } }
                type: object
        '404':
          description: 'CreekConditions not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - CreekConditions
      summary: 'Add a new CreekConditions to the store'
      operationId: addCreekConditions
      requestBody:
        $ref: '#/components/requestBodies/CreekConditions'
      responses:
        '201':
          description: 'Created CreekConditions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreekConditions'
            application/xml:
              schema:
                $ref: '#/components/schemas/CreekConditions'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/creekConditions/{id}':
    get:
      tags:
        - CreekConditions
      summary: 'Find CreekConditions by ID'
      description: 'Returns a single CreekConditions'
      operationId: getCreekConditionsById
      parameters:
        -
          name: id
          in: path
          description: 'ID of CreekConditions to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreekConditions'
            application/xml:
              schema:
                $ref: '#/components/schemas/CreekConditions'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'CreekConditions not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - CreekConditions
      summary: 'Update an existing CreekConditions'
      operationId: updateCreekConditions
      parameters:
        -
          name: id
          in: path
          description: 'CreekConditions id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/CreekConditions'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'CreekConditions not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - CreekConditions
      summary: 'Deletes a CreekConditions'
      operationId: deleteCreekConditions
      parameters:
        -
          name: id
          in: path
          description: 'CreekConditions id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /creeks:
    get:
      tags:
        - Creeks
      summary: 'Find list Creeks'
      description: 'Returns list of Creeks'
      operationId: getCreeks
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Creeks' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Creeks' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/Creeks' } }
                type: object
        '404':
          description: 'Creeks not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - Creeks
      summary: 'Add a new Creeks to the store'
      operationId: addCreeks
      requestBody:
        $ref: '#/components/requestBodies/Creeks'
      responses:
        '201':
          description: 'Created Creeks'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Creeks'
            application/xml:
              schema:
                $ref: '#/components/schemas/Creeks'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/creeks/{id}':
    get:
      tags:
        - Creeks
      summary: 'Find Creeks by ID'
      description: 'Returns a single Creeks'
      operationId: getCreeksById
      parameters:
        -
          name: id
          in: path
          description: 'ID of Creeks to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Creeks'
            application/xml:
              schema:
                $ref: '#/components/schemas/Creeks'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'Creeks not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - Creeks
      summary: 'Update an existing Creeks'
      operationId: updateCreeks
      parameters:
        -
          name: id
          in: path
          description: 'Creeks id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/Creeks'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Creeks not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - Creeks
      summary: 'Deletes a Creeks'
      operationId: deleteCreeks
      parameters:
        -
          name: id
          in: path
          description: 'Creeks id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
...
    CreekConditions:
      title: CreekConditions
      description: CreekConditions
      properties:
        creek_condition_id:
          nullable: false
          title: creek_condition_id
          description: creek_condition_id
          type: string
          format: '-'
          maxLength: 6
        observer_id:
          nullable: false
          title: observer_id
          description: observer_id
          type: string
          format: '-'
          maxLength: 6
        creek_id:
          nullable: false
          title: creek_id
          description: creek_id
          type: string
          format: '-'
          maxLength: 6
        creek_condition_description:
          nullable: false
          title: creek_condition_description
          description: creek_condition_description
          type: string
          format: '-'
        creek_condition_longitude:
          nullable: false
          title: creek_condition_longitude
          description: creek_condition_longitude
          type: string
          format: '-'
          maxLength: 100
        creek_condition_lattitude:
          nullable: false
          title: creek_condition_lattitude
          description: creek_condition_lattitude
          type: string
          format: '-'
          maxLength: 100
        creek_condition_status:
          nullable: false
          title: creek_condition_status
          description: creek_condition_status
          type: string
          format: '-'
        creek_condition_handling_status:
          nullable: false
          title: creek_condition_handling_status
          description: creek_condition_handling_status
          type: string
          format: '-'
        created_at:
          nullable: false
          title: created_at
          description: created_at
          type: string
          format: '-'
        updated_at:
          nullable: false
          title: updated_at
          description: updated_at
          type: string
          format: '-'
      type: object
    Creeks:
      title: Creeks
      description: Creeks
      properties:
        creek_id:
          nullable: false
          title: creek_id
          description: creek_id
          type: string
          format: '-'
          maxLength: 6
        river_id:
          nullable: false
          title: river_id
          description: river_id
          type: string
          format: '-'
          maxLength: 6
        creek_name:
          nullable: false
          title: creek_name
          description: creek_name
          type: string
          format: '-'
          maxLength: 50
        creek_geojson:
          nullable: false
          title: creek_geojson
          description: creek_geojson
          type: string
          format: '-'
          maxLength: 100
        creek_longitude:
          nullable: false
          title: creek_longitude
          description: creek_longitude
          type: string
          format: '-'
          maxLength: 100
        creek_lattitude:
          nullable: false
          title: creek_lattitude
          description: creek_lattitude
          type: string
          format: '-'
          maxLength: 100
        created_at:
          nullable: false
          title: created_at
          description: created_at
          type: string
          format: '-'
        updated_at:
          nullable: false
          title: updated_at
          description: updated_at
          type: string
          format: '-'
      type: object
...
  requestBodies:
...
    CreekConditions:
      description: 'CreekConditions object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/CreekConditions'
        application/xml:
          schema:
            $ref: '#/components/schemas/CreekConditions'
        application/json:
          schema:
            $ref: '#/components/schemas/CreekConditions'
    Creeks:
      description: 'Creeks object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/Creeks'
        application/xml:
          schema:
            $ref: '#/components/schemas/Creeks'
        application/json:
          schema:
            $ref: '#/components/schemas/Creeks'
...